requirements.txt
boto3==1.20.42
Flask==2.0.2
kafka==1.3.5
kafka_python==2.0.2
prometheus_client==0.11.0
PyYAML==6.0
setuptools==56.0.0
Werkzeug==2.0.2
